pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-google-community>=2.0.10",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "langgraph>=1.0.1",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.1.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
]
